fix delete files and folders
